Delete the types/category.ts file
