--- types/category.ts
@@ -1,15 +1,0 @@
-export type CategoryTheme =
-  | 'GRAY'
-  | 'RED'
-  | 'YELLOW'
-  | 'GREEN'
-  | 'BLUE'
-  | 'INDIGO'
-  | 'PURPLE'
-  | 'PINK';
-
-export type Category = {
-  id: string; // uuid
-  name: string;
-  theme: CategoryTheme;
-};
